datatype fixes and add gw630X
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@34790 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: 3f6dee07acf1902633a9db22e87c7f83ee9af81d

+++ src/linux/universal/linux-4.14/drivers/mtd/spi-nor/spi-nor.c
@@ -1491,7 +1491,7 @@ static int spi_nor_chunked_write(struct mtd_info *mtd, loff_t _to, size_t _len,
 	struct spi_nor *nor = mtd_to_spi_nor(mtd);
 	u32 addr_width = nor->addr_width + !!(nor->flags & SNOR_F_4B_EXT_ADDR);
 	int chunk_size;
-	int retlen = 0;
+	size_t retlen = 0;
 	int ret;
 
 	chunk_size = nor->chunk_size;
@@ -1534,7 +1534,7 @@ static int spi_nor_chunked_read(struct mtd_info *mtd, loff_t _from, size_t _len,
 		size_t len = min_t(int, chunk_size, _len - *_retlen);
 		u_char *buf = _buf + *_retlen;
 		loff_t from = _from + *_retlen;
-		int retlen = 0;
+		size_t retlen = 0;
 
 		ret = spi_nor_read(mtd, from, len, &retlen, buf);
 		if (ret)
